# openapi.json
{
  "components" : {
    "schemas" : {
...
      "CustomerInput" : {
        "properties" : {
          "archived" : {
            "type" : [ "boolean", "null" ]
          },
          "creditLimitMinor" : {
            "format" : "int64",
            "type" : [ "integer", "null" ]
          },
          "name" : {
            "type" : [ "string", "null" ]
          },
          "notes" : {
            "type" : [ "string", "null" ]
          },
          "phone" : {
            "type" : [ "string", "null" ]
          }
        },
        "type" : "object"
      },
      "CustomerView" : {
        "properties" : {
          "archived" : {
            "type" : "boolean"
          },
          "balanceMinor" : {
            "format" : "int64",
            "type" : "integer"
          },
          "creditLimitMinor" : {
            "format" : "int64",
            "type" : [ "integer", "null" ]
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "lastReminderAt" : {
            "format" : "date-time",
            "type" : [ "string", "null" ]
          },
          "name" : {
            "type" : [ "string", "null" ]
          },
          "notes" : {
            "type" : [ "string", "null" ]
          },
          "oldestOpenAt" : {
            "format" : "date-time",
            "type" : [ "string", "null" ]
          },
          "phone" : {
            "type" : [ "string", "null" ]
          },
          "rev" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "archived", "balanceMinor", "id", "rev" ],
        "type" : "object"
      },
...
    }
  },
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/b/{businessId}/customers/{customerId}" : {
      "get" : {
        "operationId" : "customer",
        "parameters" : [ {
          "in" : "path",
          "name" : "businessId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "customerId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "X-Member-Id",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CustomerView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "credit-controller" ]
      },
      "put" : {
        "operationId" : "upsertCustomer",
        "parameters" : [ {
          "in" : "path",
          "name" : "businessId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "customerId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "X-Member-Id",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CustomerInput"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CustomerView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "credit-controller" ]
      }
    },
...
    "/api/b/{businessId}/purchases/{purchaseId}/payments" : {
      "get" : {
        "operationId" : "payments",
        "parameters" : [ {
          "in" : "path",
          "name" : "businessId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "purchaseId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "X-Member-Id",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/SupplierPaymentView"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "stock-controller" ]
      }
    },
...
    "/api/b/{businessId}/reports/closings" : {
      "get" : {
        "operationId" : "closings",
        "parameters" : [ {
          "in" : "path",
          "name" : "businessId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "X-Member-Id",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "from",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "to",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/MemberClosings"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "report-controller" ]
      }
    },
    "/api/b/{businessId}/reports/closings.csv" : {
      "get" : {
        "operationId" : "closingsCsv",
        "parameters" : [ {
          "in" : "path",
          "name" : "businessId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "X-Member-Id",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "lang",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "from",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "to",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "text/csv" : {
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "report-controller" ]
      }
    },
    "/api/b/{businessId}/reports/daily-close" : {
      "get" : {
        "operationId" : "dailyClose",
        "parameters" : [ {
          "in" : "path",
          "name" : "businessId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "X-Member-Id",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "from",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "to",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DailyClose"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "report-controller" ]
      }
    },
    "/api/b/{businessId}/reports/expenses" : {
      "get" : {
        "operationId" : "expenses",
        "parameters" : [ {
          "in" : "path",
          "name" : "businessId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "X-Member-Id",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "from",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "to",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ExpenseReport"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "report-controller" ]
      }
    },
    "/api/b/{businessId}/reports/expenses.csv" : {
      "get" : {
        "operationId" : "expensesCsv",
        "parameters" : [ {
          "in" : "path",
          "name" : "businessId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "X-Member-Id",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "lang",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "from",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "to",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "text/csv" : {
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "report-controller" ]
      }
    },
    "/api/b/{businessId}/reports/inventory" : {
      "get" : {
        "operationId" : "inventory",
        "parameters" : [ {
          "in" : "path",
          "name" : "businessId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "X-Member-Id",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Inventory"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "report-controller" ]
      }
    },
    "/api/b/{businessId}/reports/inventory.csv" : {
      "get" : {
        "operationId" : "inventoryCsv",
        "parameters" : [ {
          "in" : "path",
          "name" : "businessId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "X-Member-Id",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "lang",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "text/csv" : {
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "report-controller" ]
      }
    },
    "/api/b/{businessId}/reports/overview" : {
      "get" : {
        "operationId" : "overview",
        "parameters" : [ {
          "in" : "path",
          "name" : "businessId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "X-Member-Id",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "from",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "to",
          "required" : false,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Overview"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "report-controller" ]
      }
    },
...
    "/api/b/{businessId}/reports/receivables" : {
      "get" : {
        "operationId" : "receivables",
        "parameters" : [ {
          "in" : "path",
          "name" : "businessId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "X-Member-Id",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Receivables"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "report-controller" ]
      }
    },
...
  },
  "servers" : [ {
    "description" : "Generated server url",
    "url" : "http://localhost"
  } ]
}